PIM-5099: Add dedicated behat context for xlsx files
