+++ features/acl/assert_acl.feature
@@ -45,7 +45,7 @@ Feature: Define user rights
       | Remove a product            | "boot" product                              | Delete                  |                          |
       | Download the product as PDF | "boot" product                              | Pdf                     |                          |
       | Remove an attribute         | "color" attribute                           | Delete                  |                          |
-      | Remove an export profile    | "footwear_option_export" export job edit    | Delete                  |                          |
+      | Remove an export profile    | "csv_footwear_option_export" export job edit    | Delete                  |                          |
       | Remove an import profile    | "csv_footwear_group_import" import job edit | Delete                  |                          |
 
   @jira https://akeneo.atlassian.net/browse/PIM-3758
